Extracted module utils and docs fragment.
